Karan-Kakkar.md
Guess The Number
This is a Number guessing game made using Python. It uses random numbers, which guesses a number between 1 and 20(the limits can be changed according to the need of the user. The user needs to answer the questions such as whether the number the user guessed is greater or smaller. Finally the program will e able to tell the numer guessed by the user. Hope you like it....