Create Heap_Sort.c
Heap sort processes the elements by creating the min heap or max heap using the elements of the given array. Min heap or max heap represents the ordering of the array in which root element represents the minimum or maximum element of the array. At each step, the root element of the heap gets deleted and stored into the sorted array and the heap will again be heapified.

The heap sort basically recursively performs two main operations.

Build a heap H, using the elements of ARR.
Repeatedly delete the root element of the heap formed in phase 1.