Improved the fact function.
I increased the range of  the variables(int) by choosing long long datatype, so the program won't have number limiting issues to a certain extent. I also made some changes  in indentation so that code is more readable and uniform. Also, now due to the changes made in the for loop of fact function it takes half as much iterations than it used to earlier. Hence, improving the code by time and range of datatype storage.