Book Allocation Problem
Given an array ‘pages’ of integer numbers, where ‘pages[i]’ represents the number of pages in the ‘i-th’ book. There are ‘m’ number of students, and the task is to allocate all the books to their students. 

Allocate books in a way such that:

1. Each student gets at least one book.

2. Each book should be allocated to a student.

3. Book allocation should be in a contiguous manner.

You have to allocate the books to ‘m’ students such that the maximum number of pages assigned to a student is minimum.

Let’s understand the problem statement through an example.

Example: 

Number of books = 4 and Number of students = 2

pages[] = { 10,20,30,40}

Important points to consider
Assign at least one book to every student, so there can’t be any allocation such that a student gets no books assigned.
While allocating the books, no book should be left out. In other words, you have to allocate each and every book given.
Allocate in a contiguous manner. Let's say you have to allocate 3 books to a student from pages[] = { 10,20,30,40}
 

Then, the possible allocations can be - {10,20,30} and {20,30,40}. You can’t allocate {10,30,40} as it is not contiguous.

All possible ways of book allocation are shown in the below figure-

The minimum of the maximum number of pages assigned = min{90,70,60} = 60. Hence, the required answer is 60.

Brute Force Approach
Let N be the number of books and M the number of students.

In the case of M > N:

Simply return -1 as the number of students is greater than the number of books available, and according to the given constraints, it is not possible to allocate at least one book to each student.

 Let’s see the algorithm for all other cases where M<=N:

We have to minimize the value of the maximum number of pages assigned to a student in an allocation. 
If the maximum number of pages assigned to a student in a book allocation is max_pages, then this implies that the number of pages assigned to every student is less than or equal to max_pages.
Let’s focus on the possible values of the number of pages that can be assigned to a student? 
The minimum value can be equal to 0 (Though not in this problem because you have to assign each student at least one book, so the number of pages can't be zero)
The maximum number of pages will be the sum of the number of pages in all books. (This will happen when you assign all the books to one student).
The range of the maximum number of pages we obtained is - (0, sum of all the values of pages array]. Open interval on 0 because at least one book needs to be assigned to every student. So, the interval we get is [1, sum of pages array].
Steps of algorithm
Iterate over all values ranging from numPages=1 to numPages=sum of pages array.
Check if it is possible to allocate the books such that the value of the number of pages assigned to any student is less than or equal to numPages in each iteration.
If the maximum number of pages can be equal to numPages, then return numPages as the answer. We don’t need to iterate further because after this the value of numPages will increase, but we are interested in finding the minimum value of the maximum number of pages that can be allocated. 
 
How to check if it is possible to allocate the books such that the maximum number of pages assigned to any student is numPages?

Initialize count of students with 1.
Keep allocating the books to a student until the sum of the pages assigned is less than numPages. 
If at any point the number of pages assigned to a student exceeds numPages, then allocate the current book to the next student and increment the count of students.
If the count of students becomes greater than M, then return false.
In the end, if the count of students is equal to M, return true.